refactor: Move the primary's messages to their own crates
