+++ consensus/src/dag.rs
@@ -1,9 +1,9 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 use crypto::traits::VerifyingKey;
-use primary::Certificate;
 use tokio::{sync::mpsc::Receiver, task::JoinHandle};
 use tracing::debug;
+use types::Certificate;
 
 /// Dag represents the pure dag that is constructed
 /// by the certificate of each round without any
