+++ primary/src/lib.rs
@@ -8,11 +8,6 @@
     rust_2021_compatibility
 )]
 
-#[macro_use]
-extern crate derive_builder;
-
-#[macro_use]
-mod error;
 mod aggregators;
 mod block_remover;
 mod block_waiter;
@@ -21,7 +16,6 @@ mod core;
 mod garbage_collector;
 mod header_waiter;
 mod helper;
-mod messages;
 mod payload_receiver;
 mod primary;
 mod proposer;
@@ -34,11 +28,7 @@ mod common;
 pub use crate::{
     block_remover::{BlockRemover, BlockRemoverCommand, DeleteBatchMessage},
     block_waiter::{BatchMessage, BlockCommand, BlockWaiter},
-    messages::{
-        Batch, BatchDigest, Certificate, CertificateDigest, Header, HeaderDigest, Transaction,
-    },
     primary::{
-        PayloadToken, Primary, PrimaryWorkerMessage, Round, WorkerPrimaryError,
-        WorkerPrimaryMessage,
+        PayloadToken, Primary, PrimaryWorkerMessage, WorkerPrimaryError, WorkerPrimaryMessage,
     },
 };
