+++ primary/src/payload_receiver.rs
@@ -5,8 +5,9 @@ use config::WorkerId;
 
 use store::Store;
 use tokio::sync::mpsc::Receiver;
+use types::BatchDigest;
 
-use crate::{messages::BatchDigest, primary::PayloadToken};
+use crate::primary::PayloadToken;
 
 /// Receives batches' digests of other authorities. These are only needed to verify incoming
 /// headers (i.e.. make sure we have their payload).
