+++ primary/src/tests/block_waiter_tests.rs
@@ -1,15 +1,11 @@
+use crate::{BatchMessage, BlockCommand, BlockWaiter};
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 use crate::{
     block_waiter::{
         BatchResult, BlockError, BlockErrorType, BlockResult, GetBlockResponse, GetBlocksResponse,
     },
-    common,
-    common::{
-        certificate, create_db_stores, fixture_batch_with_transactions, fixture_header_builder,
-        keys, resolve_name_and_committee,
-    },
-    messages, BatchDigest, BatchMessage, BlockCommand, BlockWaiter, Certificate, CertificateDigest,
+    common::create_db_stores,
     PrimaryWorkerMessage,
 };
 use bincode::deserialize;
@@ -28,6 +24,13 @@ use tokio::{
     time::{sleep, timeout, Duration},
 };
 use tokio_util::codec::{Framed, LengthDelimitedCodec};
+use types::{
+    test_utils::{
+        certificate, fixture_batch_with_transactions, fixture_header_builder,
+        fixture_header_with_payload, keys, resolve_name_and_committee,
+    },
+    Batch, BatchDigest, Certificate, CertificateDigest,
+};
 
 #[tokio::test]
 async fn test_successfully_retrieve_block() {
@@ -38,7 +41,7 @@ async fn test_successfully_retrieve_block() {
     let (name, committee) = resolve_name_and_committee(13000);
 
     // AND store certificate
-    let header = common::fixture_header_with_payload(2);
+    let header = fixture_header_with_payload(2);
     let certificate = certificate(&header);
     certificate_store
         .write(certificate.digest(), certificate.clone())
@@ -66,7 +69,7 @@ async fn test_successfully_retrieve_block() {
             batch_id,
             BatchMessage {
                 id: batch_id,
-                transactions: messages::Batch(vec![vec![10u8, 5u8, 2u8], vec![8u8, 2u8, 3u8]]),
+                transactions: Batch(vec![vec![10u8, 5u8, 2u8], vec![8u8, 2u8, 3u8]]),
             },
         );
     }
@@ -268,7 +271,7 @@ async fn test_one_pending_request_for_block_at_time() {
     let (name, committee) = resolve_name_and_committee(13002);
 
     // AND store certificate
-    let header = common::fixture_header_with_payload(2);
+    let header = fixture_header_with_payload(2);
     let certificate = certificate(&header);
     certificate_store
         .write(certificate.digest(), certificate.clone())
@@ -336,7 +339,7 @@ async fn test_unlocking_pending_get_block_request_after_response() {
     let (name, committee) = resolve_name_and_committee(13003);
 
     // AND store certificate
-    let header = common::fixture_header_with_payload(2);
+    let header = fixture_header_with_payload(2);
     let certificate = certificate(&header);
     certificate_store
         .write(certificate.digest(), certificate.clone())
@@ -395,7 +398,7 @@ async fn test_batch_timeout() {
     let (name, committee) = resolve_name_and_committee(13004);
 
     // AND store certificate
-    let header = common::fixture_header_with_payload(2);
+    let header = fixture_header_with_payload(2);
     let certificate = certificate(&header);
     certificate_store
         .write(certificate.digest(), certificate.clone())
