+++ primary/src/tests/core_tests.rs
@@ -2,15 +2,15 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 use super::*;
-use crate::{
-    common::{
-        certificate, committee, committee_with_base_port, create_db_stores, header, headers, keys,
-        listener, votes,
-    },
-    BatchDigest,
-};
+use crate::common::create_db_stores;
 use crypto::traits::KeyPair;
 use futures::future::try_join_all;
+use types::{
+    test_utils::{
+        certificate, committee, committee_with_base_port, header, headers, keys, listener, votes,
+    },
+    BatchDigest, Header, Vote,
+};
 
 use tokio::sync::mpsc::channel;
