+++ types/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "types"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+authors = ["Mysten Labs <build@mystenlabs.com>"]
+publish = false
+
+
+[dependencies]
+futures = "0.3.21"
+serde = { version = "1.0.136", features = ["derive"] }
+tokio = { version = "1.17.0", features = ["sync", "rt", "macros"] }
+tokio-util = { version = "0.7.0", features = ["codec"] }
+ed25519-dalek = "1.0.1"
+base64 = "0.13.0"
+blake2 = "0.9"
+derive_builder = "0.11.1"
+store = { git = "https://github.com/mystenlabs/mysten-infra.git", package = "typed-store", rev = "808de09203d147b43d59114b8afd9e51cbcf5778"}
+thiserror = "1.0.30"
+bincode = "1.3.3"
+bytes = "1.1.0"
+rand = "0.7.3"
+tempfile = "3.3.0"
+
+
+crypto = { path = "../crypto" }
+config = { path = "../config" }
+
+
+[features]
+default = []
+test = []
