+++ types/src/error.rs
@@ -1,7 +1,7 @@
 // Copyright (c) 2021, Facebook, Inc. and its affiliates
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
-use crate::{messages::HeaderDigest, primary::Round};
+use crate::{HeaderDigest, Round};
 use crypto::{CryptoError, Digest};
 use store::StoreError;
 use thiserror::Error;
